requirements.txt
pyTelegramBotAPI==4.12.0
transformers==4.31.0
torch==2.0.1
spacy==3.5.0
en-core-web-sm==3.5.0
de-core-news-sm==3.5.0
fr-core-news-sm==3.5.0
uk-core-news-sm==3.5.0
ru-core-news-sm==3.5.0
pl-core-news-sm==3.5.0
langdetect==1.0.9
pdfplumber==0.9.0
python-docx==0.8.11
numpy==1.24.0
PyPDF2==3.0.1